Make gallery picture edit form a child page of gallery picture modal
