--- app/routes/photos/components/GalleryPictureEditModal.tsx
@@ -1,186 +1,0 @@
-import {
-  Button,
-  ButtonGroup,
-  Flex,
-  Icon,
-  Image,
-  Modal,
-  LoadingPage,
-} from '@webkom/lego-bricks';
-import { usePreparedEffect } from '@webkom/react-prepare';
-import { Trash2 } from 'lucide-react';
-import { Field } from 'react-final-form';
-import { Link, useNavigate, useParams } from 'react-router-dom';
-import {
-  deletePicture,
-  fetchGalleryPicture,
-  updatePicture,
-} from 'app/actions/GalleryPictureActions';
-import {
-  Form,
-  TextArea,
-  SelectInput,
-  CheckBox,
-  LegoFinalForm,
-} from 'app/components/Form';
-import { SubmitButton } from 'app/components/Form/SubmitButton';
-import { selectGalleryById } from 'app/reducers/galleries';
-import { selectGalleryPictureById } from 'app/reducers/galleryPictures';
-import { useAppDispatch, useAppSelector } from 'app/store/hooks';
-import GalleryDetailsRow from './GalleryDetailsRow';
-import styles from './GalleryPictureModal.module.css';
-import type { EntityId } from '@reduxjs/toolkit';
-import type { DetailedGallery } from 'app/store/models/Gallery';
-
-type FormValues = {
-  description: string;
-  active: boolean;
-  taggees: { label: string; value: EntityId }[];
-};
-
-const TypedLegoForm = LegoFinalForm<FormValues>;
-
-const GalleryPictureEditModal = () => {
-  const { pictureId, galleryId } = useParams<{
-    pictureId: string;
-    galleryId: string;
-  }>();
-  const fetching = useAppSelector(
-    (state) => state.galleries.fetching || state.galleryPictures.fetching,
-  );
-  const picture = useAppSelector((state) =>
-    selectGalleryPictureById(state, pictureId),
-  );
-  const gallery = useAppSelector((state) =>
-    selectGalleryById<DetailedGallery>(state, galleryId),
-  );
-
-  const dispatch = useAppDispatch();
-
-  usePreparedEffect(
-    'fetchGalleryAndGalleryPicture',
-    () =>
-      galleryId &&
-      pictureId &&
-      dispatch(fetchGalleryPicture(galleryId, pictureId)),
-    [dispatch, galleryId, pictureId],
-  );
-
-  const navigate = useNavigate();
-
-  if (!gallery || !picture) {
-    return <LoadingPage loading={fetching} />;
-  }
-
-  const onSubmit = (data) => {
-    const body = {
-      id: picture.id,
-      gallery: gallery.id,
-      description: data.description,
-      active: data.active,
-      taggees: data.taggees && data.taggees.map((taggee) => taggee.value),
-    };
-    dispatch(updatePicture(body)).then(() => {
-      navigate(`/photos/${gallery.id}/picture/${picture.id}`);
-    });
-  };
-
-  const initialValues = {
-    ...picture,
-    taggees:
-      picture &&
-      picture.taggees.map((taggee) => ({
-        value: taggee.id,
-        label: taggee.fullName,
-      })),
-  };
-
-  return (
-    <Modal
-      onOpenChange={(open) => !open && navigate(`/photos/${gallery.id}`)}
-      isOpen
-      isDismissable={false}
-      contentClassName={styles.content}
-    >
-      <Flex column gap="var(--spacing-md)">
-        <Flex width="100%" justifyContent="space-between" alignItems="center">
-          <Flex justifyContent="space-between">
-            <Image
-              className={styles.galleryThumbnail}
-              alt="Forsidebilde til album"
-              src={gallery.cover?.thumbnail ?? ''}
-            />
-
-            <Flex column justifyContent="space-around">
-              <h5>
-                <Link to={`/photos/${gallery.id}`}>{gallery.title}</Link>
-              </h5>
-              <GalleryDetailsRow small gallery={gallery} />
-            </Flex>
-          </Flex>
-        </Flex>
-
-        <Flex className={styles.pictureContainer}>
-          <Image
-            key={picture.id}
-            src={picture.file}
-            placeholder={picture.thumbnail}
-            alt={picture.description}
-            className={styles.picture}
-          />
-        </Flex>
-
-        <TypedLegoForm onSubmit={onSubmit} initialValues={initialValues}>
-          {({ handleSubmit }) => (
-            <Form onSubmit={handleSubmit}>
-              <Field
-                label="Bildebeskrivelse"
-                placeholder="Beskrivelse"
-                name="description"
-                component={TextArea.Field}
-              />
-              <Field
-                label="Synlig for alle brukere"
-                name="active"
-                type="checkbox"
-                component={CheckBox.Field}
-              />
-              <Field
-                label="Tagg brukere"
-                name="taggees"
-                filter={['users.user']}
-                placeholder="Skriv inn navn på brukere i bildet"
-                component={SelectInput.AutocompleteField}
-                isMulti
-              />
-              <ButtonGroup>
-                <Button
-                  flat
-                  onPress={() => {
-                    navigate(`/photos/${gallery.id}/picture/${picture.id}`);
-                  }}
-                >
-                  Avbryt
-                </Button>
-                <SubmitButton>Lagre</SubmitButton>
-                <Button
-                  danger
-                  onPress={() =>
-                    dispatch(deletePicture(gallery.id, picture.id)).then(() => {
-                      navigate(`/photos/${gallery.id}`);
-                    })
-                  }
-                >
-                  <Icon iconNode={<Trash2 />} size={19} />
-                  Slett bildet
-                </Button>
-              </ButtonGroup>
-            </Form>
-          )}
-        </TypedLegoForm>
-      </Flex>
-    </Modal>
-  );
-};
-
-export default GalleryPictureEditModal;
